+++ launchable/commands/subset.py
@@ -1,7 +1,7 @@
 import click
 import os
 import sys
-from os.path import *
+from os.path import join, relpath, normpath
 import glob
 from typing import Callable, Union, Optional
 from ..utils.click import PERCENTAGE, DURATION
@@ -10,7 +10,7 @@ from ..utils.http_client import LaunchableClient
 from ..testpath import TestPath
 from .helper import find_or_create_session
 from ..utils.click import KeyValueType
-
+from .helper import TestPathWriter
 # TODO: rename files and function accordingly once the PR landscape
 
 
@@ -245,48 +245,3 @@ def subset(context, target, session: Optional[str], base_path: Optional[str], bu
                 self.print(output)
 
     context.obj = Optimize()
-
-
-class TestPathWriter(object):
-    base_path = None
-
-    def __init__(self):
-        self._formatter = TestPathWriter.default_formatter
-        self._separator = "\n"
-
-    @classmethod
-    def default_formatter(cls, x: TestPath):
-        """default formatter that's in line with to_test_path(str)"""
-        file_name = x[0]['name']
-        if cls.base_path:
-            # default behavior consistent with default_path_builder's relative path handling
-            file_name = join(cls.base_path, file_name)
-        return file_name
-
-    @property
-    def formatter(self) -> Callable[[TestPath], str]:
-        """
-        This function, if supplied, is used to format test names
-        from the format Launchable uses to the format test runners expect.
-        """
-        return self._formatter
-
-    @formatter.setter
-    def formatter(self, v: Callable[[TestPath], str]):
-        self._formatter = v
-
-    @property
-    def separator(self) -> str:
-        return self._separator
-
-    @separator.setter
-    def separator(self, s: str):
-        self._separator = s
-
-    def write_file(self, file: str, test_paths: []):
-        open(file, "w+", encoding="utf-8").write(
-            self.separator.join(self.formatter(x=t) for t in test_paths))
-
-    def print(self, test_paths: []):
-        click.echo(self.separator.join(self.formatter(x=t)
-                                       for t in test_paths))
